QNofcustomdata/requirements.txt
langchain
google-generativeai
streamlit
google-cloud-translate
docx2txt

